refactor: Improve modals handling for `SidebarActions` component
